rename ClassConstantReplacerRector to RenameClassConstantRector
