--- tests/Rector/Constant/ClassConstantReplacerRector/Source/LocalFormEvents.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\Constant\ClassConstantReplacerRector\Source;
-
-final class LocalFormEvents
-{
-
-}
