+++ tests/Rector/Constant/RenameClassConstantRector/Source/DifferentClass.php
@@ -1,6 +1,6 @@
 <?php declare (strict_types=1);
 
-namespace Rector\Tests\Rector\Constant\ClassConstantReplacerRector\Source;
+namespace Rector\Tests\Rector\Constant\RenameClassConstantRector\Source;
 
 final class DifferentClass
 {
