+++ tests/Rector/Constant/RenameClassConstantRector/Source/LocalFormEvents.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Constant\RenameClassConstantRector\Source;
+
+final class LocalFormEvents
+{
+
+}
